+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.1
 
 - Added interface `WithDimension` to `PDimension` and `PCanvas`.
+- Added `PCanvas.info`.
 - README.md: added usage example.
 
 ## 1.0.0